requirements.txt
fastapi
uvicorn
sqlalchemy
gpxpy
geopy
python-multipart
jinja2
passlib
python-jose
aiofiles
requests
markdown
google-generativeai
python-dotenv
python-slugify
psycopg2-binary
GeoAlchemy2
